query_eval: split eval integration tests into one file per query-node type
mod.rs is getting quite big, let's split it.
